requirements.txt
numpy>=1.26.0
pandas>=2.2.0
seaborn>=0.13.0
matplotlib>=3.8.0
rdkit==2025.09.3
scikit-learn>=1.8.0
tqdm>=4.66.0
transformers>=4.40.0
pytorch-lightning>=2.2.0
torchmetrics>=1.4.0
peft>=0.10.0
fastapi[standard]>=0.116.1
uvicorn[standard]>=0.30.1
pydantic-settings>=2.10.1
sqlalchemy>=2.0.45
alembic>=1.17.2
pyjwt>=2.10.1
tzdata>=2025.2
pwdlib>=0.3.0
aiosqlite>=0.21.0
python-jose>=3.5.0 
optuna>=4.6.0 
xgboost>=3.1.2
skl2onnx>=1.19.1
onnxruntime>=1.23.2
onnxmltools>=1.14.0
lightgbm>=4.6.0
streamlit>=1.52.1
torch>=2.9